+++ lib/syscalls/chdir.c
@@ -1,5 +1,5 @@
 #include "syscall_preamble.h"
-int chdir(const char *path) {
+long chdir(const char *path) {
   int ret;
   syscall_output()
   syscall_input(chdir, "D"(path))
