+++ lib/syscalls/rmdir.c
@@ -1,5 +1,5 @@
 #include "syscall_preamble.h"
-int rmdir(const char *path) {
+long rmdir(const char *path) {
   int ret;
   syscall_output()
   syscall_input(rmdir, "D"(path))
